Only reboot on rollback if needed
If mender is configured to boot from the active partition,
then no rollback is necessary.
Prior to this change, when rolling back prior to the post install
reboot, the RollbackReboot state is entered, causing the system
to reboot. It is expected that the system should reboot to
change paritions, but if the active partition matches the current
one, then there is not reason to reboot.